BROKEN: In progress: Fix all errors apart from SpecifySeverity & converting to new methods
